Add DeleteRangeAsync overload for long IDs in repository
Added DeleteRangeAsync(List<long> ids) to GenericRepository and IGenericRepository to support entities with long primary keys, such as RawRFIDReading. Updated RFIDImportService to use the new overload, ensuring type safety and avoiding casting issues. Improved LINQ queries by using EF.Property to eliminate reflection.

+++ Runnatics/src/Runnatics.Services/RFIDImportService.cs
@@ -1967,8 +1967,8 @@ namespace Runnatics.Services
                 {
                     if (readings.Count > 0)
                     {
-                        // RawRFIDReading.Id is long, not int
-                        var readingIdsToDelete = readings.Select(r => (int)r.Id).ToList();
+                        // Use long overload for RawRFIDReading.Id
+                        var readingIdsToDelete = readings.Select(r => r.Id).ToList();
                         await readingRepo.DeleteRangeAsync(readingIdsToDelete);
                     }
                     if (batches.Count > 0)
